requirements.txt
Flask
Flask[async]
Flask-Cors
youtube_transcript_api
google-generativeai
python-dotenv
pathlib
streamlit